requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asyncio-mqtt==0.12.1
    # via -r requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via
    #   kubernetes
    #   requests
charset-normalizer==2.1.0
    # via requests
google-auth==2.9.0
    # via kubernetes
idna==3.3
    # via requests
kubernetes==19.15.0
    # via -r requirements.in
oauthlib==3.2.0
    # via requests-oauthlib
paho-mqtt==1.6.1
    # via asyncio-mqtt
protobuf==3.20.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
python-dateutil==2.8.2
    # via kubernetes
python-dotenv==0.20.0
    # via -r requirements.in
pyyaml==6.0
    # via kubernetes
requests==2.28.1
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
urllib3==1.26.10
    # via
    #   kubernetes
    #   requests
websocket-client==1.3.3
    # via kubernetes

# The following packages are considered to be unsafe in a requirements file:
# setuptools
